requirements.txt
lxml==4.3.0
python-dateutil==2.8.0
requests==2.21.0
